my_wedding_planner/requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.6
Pillow==8.3.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
